NEM_SLICE: Better diagnostics for memory errors in find_surnd_elem routine

+++ packages/seacas/applications/nem_slice/elb.h
@@ -243,7 +243,7 @@ struct Sphere_Info
 template <typename INT> struct Graph_Description
 {
   size_t                        nadj;
-  size_t                        max_nsur;
+  int                           max_nsur;
   std::vector<INT>              adj;
   std::vector<INT>              start;
   std::vector<std::vector<INT>> sur_elem;
